Cleaned up boxplots
Cleaned up boxplots and added the overall dataset to account for. Interesting data that conflicts with each other on overall to 10 vs county top 10. We probably need to deep dive that more as well as deep dive into why asian and pacific islander are so low compared to everyone else